+++ packages/editor/src/browser/editor-frontend-module.ts
@@ -33,6 +33,8 @@ import { NavigationLocationSimilarity } from './navigation/navigation-location-s
 import { EditorVariableContribution } from './editor-variable-contribution';
 import { SemanticHighlightingService } from './semantic-highlight/semantic-highlighting-service';
 import { EditorQuickOpenService } from './editor-quick-open-service';
+import { TextEditor } from './editor';
+import { BreadcrumbsRenderer, BreadcrumbsRendererFactory } from './breadcrumbs/breadcrumbs-renderer';
 
 export default new ContainerModule(bind => {
     bindEditorPreferences(bind);
@@ -74,4 +76,13 @@ export default new ContainerModule(bind => {
     bind(ActiveEditorAccess).toSelf().inSingletonScope();
     bind(EditorAccess).to(CurrentEditorAccess).inSingletonScope().whenTargetNamed(EditorAccess.CURRENT);
     bind(EditorAccess).to(ActiveEditorAccess).inSingletonScope().whenTargetNamed(EditorAccess.ACTIVE);
+
+    bind(BreadcrumbsRendererFactory).toFactory(ctx =>
+        (editor: TextEditor) => {
+            const childContainer = ctx.container.createChild();
+            childContainer.bind('TextEditor').toConstantValue(editor);
+            childContainer.bind(BreadcrumbsRenderer).toSelf();
+            return childContainer.get(BreadcrumbsRenderer);
+        }
+    );
 });